+++ components/support/nimbus-fml/src/backends/frontend_manifest.rs
@@ -140,10 +140,10 @@ impl From<PropDef> for FieldBody {
 }
 
 impl From<PropDef> for FeatureFieldBody {
-    #[allow(deprecated)]
     fn from(value: PropDef) -> Self {
         Self {
-            pref: value.pref.clone(),
+            pref_key: value.pref_key.clone(),
+            gecko_pref: value.gecko_pref.clone(),
             string_alias: value.string_alias.as_ref().map(TypeRef::to_string),
             field: value.into(),
         }