+++ components/support/nimbus-fml/src/backends/kotlin/templates/FeatureTemplate.kt
@@ -11,7 +11,10 @@ public class {{ inner.name()|class_name }}  {% call kt::render_constructor() %} 
             listOf(
             {%- for p in inner.props() %}
             {%- if p.has_prefs() %}
-                {{ p.pref().unwrap().key()|quoted }},
+                {{ p.pref_key().unwrap()|quoted }},
+            {%- endif %}
+            {%- if p.has_gecko_prefs() %}
+                {{ p.gecko_pref().unwrap().pref()|quoted }},
             {%- endif %}
             {%- endfor %}
             ).any { prefs.contains(it) }